[#110]; refactor: 클로드 코드 리뷰 반영
